+++ server/src/main/java/org/apache/iotdb/db/conf/IoTDBDescriptor.java
@@ -151,7 +151,7 @@ public class IoTDBDescriptor {
       conf.setRpcThriftCompressionEnable(
           Boolean.parseBoolean(
               properties.getProperty(
-                  "rpc_thrift_compression_enable",
+                  "rpc_compression_enable",
                   Boolean.toString(conf.isRpcThriftCompressionEnable()))));
 
       conf.setRpcAdvancedCompressionEnable(
@@ -619,7 +619,7 @@ public class IoTDBDescriptor {
       conf.setThriftMaxFrameSize(
           Integer.parseInt(
               properties.getProperty(
-                  "thrift_max_frame_size", String.valueOf(conf.getThriftMaxFrameSize()))));
+                  "rpc_max_frame_size", String.valueOf(conf.getThriftMaxFrameSize()))));
 
       if (conf.getThriftMaxFrameSize() < IoTDBConstant.LEFT_SIZE_IN_REQUEST * 2) {
         conf.setThriftMaxFrameSize(IoTDBConstant.LEFT_SIZE_IN_REQUEST * 2);
@@ -628,7 +628,7 @@ public class IoTDBDescriptor {
       conf.setThriftDefaultBufferSize(
           Integer.parseInt(
               properties.getProperty(
-                  "thrift_init_buffer_size", String.valueOf(conf.getThriftDefaultBufferSize()))));
+                  "rpc_init_buffer_size", String.valueOf(conf.getThriftDefaultBufferSize()))));
 
       conf.setFrequencyIntervalInMinute(
           Integer.parseInt(